Fixing up Hydra template file and Python exporter.
The template file was just cosmetic changes of labels and
descriptions. The Python exporter was changed to fix outputting
body forces for unassigned body forces (i.e. not assigned to a domain
set). It will get written out with the -i id if no other body
forces of the same type exist that are assigned to a domain set.